move Patcher, PatchTransformation and PatchException to api package
