+++ src/main/java/org/betonquest/betonquest/api/config/PatchException.java
@@ -1,4 +1,4 @@
-package org.betonquest.betonquest.modules.config;
+package org.betonquest.betonquest.api.config;
 
 import java.io.Serial;
